Use the install-llvm-action action for LLVM version 18 and below, and the official apt packages for 19 and above
KyleMayes/install-llvm-action uses the semi-official binaries uploaded to the LLVM github releases. For LLVM 19, these are built with LTO, which cannot be linked against by the standard linkers (and broke CI for LLVM 19).

To work around this, we can use the github action to install older versions of LLVM, but use the official apt packages for newer versions. Ubuntu 24.04 (noble) has packages starting at LLVM 17, so this should be good for a while.

This is pretty much what Inkwell seems to be doing as well: https://github.com/TheDan64/inkwell/pull/557